+++ src/common/models/notifications/subs/options/sms-options.model.ts
@@ -1,4 +1,4 @@
-import { SmsFunctionality } from 'common/enums';
+import { SmsFunctionality } from '../../../../enums';
 import { Options } from './base-options.model';
 
 export class SmsOptions extends Options {