+++ src/services/notifications/notifier.service.ts
@@ -1,6 +1,6 @@
 import { AxiosPromise, AxiosRequestConfig, CreateAxiosDefaults } from 'axios';
 import { BaseService } from '../core';
-import { Notifier, NotifierModel } from 'common';
+import { Notifier, NotifierModel } from '../../common';
 export class NotifierService extends BaseService<NotifierModel, Notifier> {
   constructor(protected readonly path: string, protected readonly options?: CreateAxiosDefaults) {
     super(path, options);