+++ src/services/notifications/providers.service.ts
@@ -1,6 +1,6 @@
 import { CreateAxiosDefaults } from 'axios';
 import { BaseService } from '../core';
-import { Provider, ProviderModel } from 'common';
+import { Provider, ProviderModel } from '../../common';
 export class ProvidersService extends BaseService<ProviderModel, Provider> {
   constructor(protected readonly path: string, protected readonly options?: CreateAxiosDefaults) {
     super(path, options);