+++ src/services/notifications/templates.service.ts
@@ -1,6 +1,6 @@
 import { CreateAxiosDefaults } from 'axios';
 import { BaseService } from '../core';
-import { Template, TemplateModel } from 'common';
+import { Template, TemplateModel } from '../../common';
 export class TemplatesService extends BaseService<TemplateModel, Template> {
   constructor(protected readonly path: string, protected readonly options?: CreateAxiosDefaults) {
     super(path, options);